Updated dist to work with decimals

+++ ann/lsh/modules/EuclidianHashing.h
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <fstream>
 #include <vector>
-#include <random>
+
 #include <iomanip>
 #include <cmath>
 #include <map>
@@ -51,7 +51,7 @@ void Euclidian_Hash_Tables_Finalization(int L); /* Finilization of Hash tables *
 void Euclidian_LSH_File(int L,int k, int N); /* Calls the appropriate functions to perform LSH queries */
 void Euclidian_LSH_File_With_Prints(int L,int k, int N); /* Calls the appropriate functions to perform LSH queries */
 
-long long int calcute_euclidian_distance(int input_line, int query_line); /* Calculates euclidian distance between a and b */
+long double calcute_euclidian_distance(int input_line, int query_line); /* Calculates euclidian distance between a and b */
 
 void Euclidian_Full_Search_NN(int query_line); /* Brute Force Nearest Neighbour Search using Euclidian metric */
